+++ src/main/kotlin/com/github/hotm/world/auranet/RenderedDependableAuraNode.kt
@@ -6,7 +6,7 @@ import java.util.stream.Stream
 interface RenderedDependableAuraNode : DependableAuraNode {
     fun getChildrenForRender(): Stream<BlockPos>
 
-    fun getSuppliedAuraForRender(pos: BlockPos): Int
+    fun getSuppliedAuraForRender(pos: BlockPos): Float
 
     fun getCrownRollSpeed(pos: BlockPos): Float
 }
